test(channels): fix presentation regression CI gates
