+++ extensions/discord/src/actions/handle-action.test.ts
@@ -1,53 +1,19 @@
 // Discord tests cover handle action plugin behavior.
 import { expectDefined } from "@openclaw/normalization-core";
-import type { OpenClawConfig } from "openclaw/plugin-sdk/config-contracts";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-
-const runtimeModule = await import("./runtime.js");
-const handleDiscordActionMock = vi
-  .spyOn(runtimeModule, "handleDiscordAction")
-  .mockResolvedValue({ content: [], details: { ok: true } });
-const { handleDiscordMessageAction } = await import("./handle-action.js");
+import {
+  defaultActionOptions,
+  discordConfig,
+  expectDiscordActionCall,
+  handleDiscordActionMock,
+  handleDiscordMessageAction,
+} from "./handle-action.test-support.js";
 const {
   beginDiscordActiveTurnThreadRoute,
   notifyDiscordActiveTurnThreadCreated,
   notifyDiscordActiveTurnThreadReplyDelivered,
 } = await import("../active-turn-thread-route.js");
 const { discordInboundEventDelivery } = await import("../inbound-event-delivery.js");
-
-function discordConfig(actions?: Record<string, boolean>): OpenClawConfig {
-  return {
-    channels: { discord: { token: "tok", ...(actions ? { actions } : {}) } },
-  } as OpenClawConfig;
-}
-
-function defaultActionOptions() {
-  return {
-    mediaAccess: undefined,
-    mediaLocalRoots: undefined,
-    mediaReadFile: undefined,
-  };
-}
-
-function expectDiscordActionCall(params: {
-  payload: unknown;
-  cfg: OpenClawConfig;
-  options?: unknown;
-}) {
-  expect(handleDiscordActionMock).toHaveBeenCalledTimes(1);
-  const [call] = handleDiscordActionMock.mock.calls;
-  if (!call) {
-    throw new Error("expected Discord action call");
-  }
-  const [payload, cfg, options] = call;
-  expect(payload).toEqual(params.payload);
-  expect(cfg).toBe(params.cfg);
-  if ("options" in params) {
-    expect(options).toEqual(params.options);
-  } else {
-    expect(options).toBeUndefined();
-  }
-}
 
 describe("handleDiscordMessageAction", () => {
   beforeEach(() => {
@@ -883,97 +849,6 @@ describe("handleDiscordMessageAction", () => {
       cfg,
       defaultActionOptions(),
     );
-  });
-
-  it("downgrades chart-only presentations to Discord component text", async () => {
-    const cfg = discordConfig();
-
-    await handleDiscordMessageAction({
-      action: "send",
-      params: {
-        to: "channel:123",
-        presentation: {
-          blocks: [
-            {
-              type: "chart",
-              chartType: "bar",
-              title: "Revenue",
-              categories: ["Q1", "Q2"],
-              series: [{ name: "USD", values: [12, 18] }],
-            },
-          ],
-        },
-      },
-      cfg,
-    });
-
-    expectDiscordActionCall({
-      payload: {
-        action: "sendMessage",
-        accountId: undefined,
-        to: "channel:123",
-        content: "",
-        mediaUrl: undefined,
-        filename: undefined,
-        replyTo: undefined,
-        components: {
-          blocks: [
-            {
-              type: "text",
-              text: "-# Revenue (bar chart)\n- USD: Q1: 12; Q2: 18",
-            },
-          ],
-        },
-        embeds: undefined,
-        asVoice: false,
-        silent: false,
-        __sessionKey: undefined,
-        __agentId: undefined,
-      },
-      cfg,
-      options: defaultActionOptions(),
-    });
-  });
-
-  it("downgrades oversized table presentations to complete text", async () => {
-    const cfg = discordConfig();
-    const authoredText = `${"x".repeat(1997)}AUTHORED_TAIL`;
-
-    await handleDiscordMessageAction({
-      action: "send",
-      params: {
-        to: "channel:123",
-        presentation: {
-          title: authoredText,
-          blocks: [
-            { type: "text", text: authoredText },
-            { type: "context", text: authoredText },
-            {
-              type: "buttons",
-              buttons: [{ label: `${"x".repeat(80)}LABEL_TAIL`, value: "choice" }],
-            },
-            {
-              type: "table",
-              caption: "Large pipeline",
-              headers: ["Account", "Stage"],
-              rows: Array.from({ length: 900 }, (_entry, index) => [
-                `account-${String(index)}-${"x".repeat(80)}`,
-                "Review",
-              ]),
-            },
-          ],
-        },
-      },
-      cfg,
-    });
-
-    const [call] = handleDiscordActionMock.mock.calls;
-    const payload = call?.[0] as Record<string, unknown> | undefined;
-    expect(payload?.components).toBeUndefined();
-    expect(payload?.content).toEqual(expect.stringContaining("account-0-"));
-    expect(payload?.content).toEqual(expect.stringContaining("account-899-"));
-    expect(String(payload?.content).split("AUTHORED_TAIL")).toHaveLength(4);
-    expect(payload?.content).toEqual(expect.stringContaining("LABEL_TAIL"));
   });
 
   it("does not use another provider's current target for Discord sends", async () => {
